requirements.txt
django==2.0.6
gunicorn==19.8.1
celery==4.2.0
djangorestframework==3.8.2
django-rest-auth==0.9.3